pyproject.toml
[tool.poetry]
name = "dropyacht"
version = "0.1.0"
description = ""
authors = ["Satvik Thakral <satvik.thakral@tothenew.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.5"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
